refactor : 필드의 데이터 타입변경, Cart패키지의 productEntity와 productRepository삭제

+++ src/main/java/com/jointpurchases/domain/order/model/dto/CancelOrder.java
@@ -17,10 +17,10 @@ public class CancelOrder {
     public static class Response {
         private String email;
         private Long orderId;
-        private Long money;
+        private Integer money;
         private String payment;
 
-        public static Response from(OrderEntity order, Long money) {
+        public static Response from(OrderEntity order, Integer money) {
             return Response.builder()
                     .email(order.getMemberEntity().getEmail())
                     .orderId(order.getOrderId())
