+++ src/main/java/com/jointpurchases/domain/order/model/dto/CreateOrder.java
@@ -11,7 +11,7 @@ public class CreateOrder {
     @Getter
     public static class Request {
         private String email;
-        private Long money;
+        private Integer money;
         private String address;
     }
